introupdate
made a few small corrections.  Removed table at end of page, as individuals will find more comprehensive documentation on the platforms' website, and because the focus of this lesson is not using multiple platforms.  Replaced with a table explaining data types more generally and how they are represented in the platform used in this lesson.