Change Challenge exercise related to FIltering
I changed the "has more than 4 citations" to "has more than 8 citations" because the dataset only has articles with "computer" that have 8 and 9 citations. if we edit this to have the citation count be greater than 8, then there will be a reason to add the the second filter. (The dataset returns 17 rows for the first condition (Subjects LIKE '%computer%'), and 11 rows when you add the second condition (Citation_Count > 8)